backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
# Use mysqlclient on non-Windows; PyMySQL on Windows to avoid build tools
mysqlclient==2.2.4; platform_system != 'Windows'
PyMySQL==1.1.1; platform_system == 'Windows'
django-cors-headers==4.4.0
Pillow==10.4.0
reportlab==4.2.2
